pyproject.toml
[project]
name = "request-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "fastembed>=0.7.4",
    "flask>=3.1.2",
    "gunicorn>=25.1.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
]
